cleaned things up so system supports TeensyOS properly
